feat(work): update complete feature - complete feature에서 불필요한 모델 및 클래스를 제거합니다. - 필요한 데이터들을 사용하여 새롭게 라우트와 뷰모델을 정의합니다.

+++ feature/work/common/src/main/kotlin/com/gradation/lift/feature/work/common/data/WorkSharedViewModel.kt
@@ -13,7 +13,7 @@ class WorkSharedViewModel @Inject constructor(
 
     val historyRoutineList: MutableStateFlow<List<CreateHistoryRoutine>> = MutableStateFlow(emptyList())
     val historyWorkRestTime : MutableStateFlow<WorkRestTime> = MutableStateFlow(WorkRestTime())
-    val historyProgress: MutableStateFlow<Int> = MutableStateFlow(0)
+    val historyProgress: MutableStateFlow<Float> = MutableStateFlow(0f)
 
     val setHistoryRoutineList: (List<CreateHistoryRoutine>) -> Unit = {
         historyRoutineList.value =it
@@ -21,7 +21,7 @@ class WorkSharedViewModel @Inject constructor(
     val setHistoryWorkRestTime: (WorkRestTime) -> Unit = {
         historyWorkRestTime.value =it
     }
-    val setHistoryProgress: (Int) -> Unit = {
+    val setHistoryProgress: (Float) -> Unit = {
         historyProgress.value =it
     }
 }
